Updated the script to now read route definitions from an external CSV file.
Format for CSV file is:-
 ';' as a delimiter
 Sections are:
name - name of the route
dir1 - name of primary direction of route (e.g. 'City')
dir2 - name of secondary direction of route (e.g. 'Craigieburn')
segments - list of segments that make up the route, with , between each,
  e.g. '0,1,2'

Added a sample input of this sort for Craigieburn line.