Refactored much of the functionality of create_route_defs_csv.py into route_segs.py - and changed so it also uses the Seg_Reference basic class to manage segments in-memory, rather than a tuple like it did previously.
(This unifies quite a bit the segment processing that's also done in the
create_network_topology_segments.py function with this group. And should make
it easier to do more topology processing in future).

Also, some minor cleanups to get_network_stats.py to use this new interface,
and not access the route shapefiles as much.